@@ -19,14 +19,18 @@
     "require": {
         "php": ">=8.1",
         "ext-mbstring": "*",
-        "cocur/slugify": "^4.0",
+        "ext-ctype": "*",
+        "ext-libxml": "*",
+        "ext-simplexml": "*",
+        "ext-zlib": "*",
+        "cocur/slugify": "^4.1",
         "mrclay/minify": "^3.0",
-        "nesbot/carbon": "^2.41",
+        "nesbot/carbon": "^2.60",
         "psr/container": "^2",
         "psr/log": "^3",
         "qubus/exception": "^2.0",
         "qubus/inheritance": "^2.0",
-        "symfony/var-dumper": "^5.3"
+        "symfony/var-dumper": "^5.4"
     },
     "require-dev": {
         "laminas/laminas-coding-standard": "^2.0",
